+++ services/signer/standard/signgeneric.go
@@ -22,6 +22,9 @@ import (
 	"github.com/attestantio/dirk/rules"
 	"github.com/attestantio/dirk/services/checker"
 	"github.com/attestantio/dirk/services/ruler"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/codes"
 )
 
 // SignGeneric signs generic data.
@@ -35,12 +38,15 @@ func (s *Service) SignGeneric(
 	core.Result,
 	[]byte,
 ) {
+	ctx, span := otel.Tracer("attestantio.dirk.services.signer.standard").Start(ctx, "SignGeneric")
+	defer span.End()
 	started := time.Now()
 
 	if credentials == nil {
 		log.Error().Msg("No credentials supplied")
 		return core.ResultFailed, nil
 	}
+	span.SetAttributes(attribute.String("client", credentials.Client))
 
 	log := log.With().
 		Str("request_id", credentials.RequestID).
@@ -52,16 +58,19 @@ func (s *Service) SignGeneric(
 	// Check input.
 	if data == nil {
 		log.Warn().Str("result", "denied").Msg("Request empty")
+		span.SetStatus(codes.Error, "Request empty")
 		s.monitor.SignCompleted(started, "generic", core.ResultDenied)
 		return core.ResultDenied, nil
 	}
 	if data.Data == nil {
 		log.Warn().Str("result", "denied").Msg("Request missing data")
+		span.SetStatus(codes.Error, "Request missing data")
 		s.monitor.SignCompleted(started, "generic", core.ResultDenied)
 		return core.ResultDenied, nil
 	}
 	if data.Domain == nil {
 		log.Warn().Str("result", "denied").Msg("Request missing domain")
+		span.SetStatus(codes.Error, "Request missing domain")
 		s.monitor.SignCompleted(started, "generic", core.ResultDenied)
 		return core.ResultDenied, nil
 	}
@@ -81,9 +90,11 @@ func (s *Service) SignGeneric(
 	wallet, account, checkRes := s.preCheck(ctx, credentials, accountName, pubKey, ruler.ActionSign)
 	if checkRes != core.ResultSucceeded {
 		s.monitor.SignCompleted(started, "generic", checkRes)
+		span.SetStatus(codes.Ok, "")
 		return checkRes, nil
 	}
 	accountName = fmt.Sprintf("%s/%s", wallet.Name(), account.Name())
+	span.SetAttributes(attribute.String("account", accountName))
 	log = log.With().Str("account", accountName).Logger()
 
 	// Confirm approval via rules.
@@ -99,10 +110,12 @@ func (s *Service) SignGeneric(
 	switch results[0] {
 	case rules.DENIED:
 		s.monitor.SignCompleted(started, "generic", core.ResultDenied)
+		span.SetStatus(codes.Ok, "")
 		log.Debug().Str("result", "denied").Msg("Denied by rules")
 		return core.ResultDenied, nil
 	case rules.FAILED:
 		s.monitor.SignCompleted(started, "generic", core.ResultFailed)
+		span.SetStatus(codes.Ok, "")
 		log.Error().Str("result", "failed").Msg("Rules check failed")
 		return core.ResultFailed, nil
 	}
@@ -110,6 +123,7 @@ func (s *Service) SignGeneric(
 	signingRoot, err := generateSigningRoot(ctx, data.Data, data.Domain)
 	if err != nil {
 		log.Error().Err(err).Str("result", "failed").Msg("Failed to generate signing root")
+		span.SetStatus(codes.Error, "Failed to generate signing root")
 		s.monitor.SignCompleted(started, "generic", core.ResultFailed)
 		return core.ResultFailed, nil
 	}
@@ -118,11 +132,13 @@ func (s *Service) SignGeneric(
 	signature, err := signRoot(ctx, account, signingRoot[:])
 	if err != nil {
 		log.Error().Err(err).Str("result", "failed").Msg("Failed to sign")
+		span.SetStatus(codes.Error, "Failed to sign")
 		s.monitor.SignCompleted(started, "generic", core.ResultFailed)
 		return core.ResultFailed, nil
 	}
 
 	log.Trace().Str("result", "succeeded").Msg("Success")
+	span.SetStatus(codes.Ok, "")
 	s.monitor.SignCompleted(started, "generic", core.ResultSucceeded)
 	return core.ResultSucceeded, signature
 }