+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - provide clearer error message on configuration issues
   - use bytes rather than string on slashing protection import
   - provide full information on the data being signed with trace logging
+  - update tracing implementation to use opentelemetry
 
 # Version 1.1.0
   - garbage collect the slashing database on startup to reduce on-disk size